refactor : 레이어 역참조 문제 해결
