+++ lime-api/src/main/java/com/programmers/lime/domains/auth/application/OAuthUserService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.programmers.lime.domains.auth.api.dto.KakaoMemberResponse;
-import com.programmers.lime.domains.member.api.dto.response.MemberLoginResponse;
+import com.programmers.lime.domains.auth.application.dto.MemberLoginServiceResponse;
 import com.programmers.lime.domains.member.domain.Member;
 import com.programmers.lime.domains.member.domain.vo.SocialType;
 import com.programmers.lime.domains.member.implementation.MemberAppender;
@@ -23,7 +23,7 @@ public class OAuthUserService {
 	private final JwtService jwtService;
 
 	@Transactional
-	public MemberLoginResponse login(final String code) {
+	public MemberLoginServiceResponse login(final String code) {
 		String kakaoAccessToken = kakaoOAuthClient.getAccessToken(code);
 		KakaoMemberResponse response = kakaoOAuthClient.getMemberInfo(kakaoAccessToken);
 
@@ -35,7 +35,7 @@ public class OAuthUserService {
 		String accessToken = jwtService.generateAccessToken(String.valueOf(foundMember.getId()));
 		String refreshToken = jwtService.generateRefreshToken();
 
-		return MemberLoginResponse.from(foundMember, accessToken, refreshToken);
+		return MemberLoginServiceResponse.from(foundMember, accessToken, refreshToken);
 	}
 
 	private Member saveMember(final KakaoMemberResponse response) {
